analysis/advanced_analyzer.py
import filetype
import matplotlib.pyplot as plt

class AdvancedAnalyzer:
    def analyze_signature(self, file_path):
        kind = filetype.guess(file_path)
        if kind:
            return f"Extension: {kind.extension}, MIME: {kind.mime}"
        else:
            return "Unknown file type"

    def plot_timeline(self, timestamps):
        # Placeholder: implement timeline visualization
        pass

    def data_carving(self, image_path):
        # Placeholder: implement data carving logic
        pass
