requirements.txt
pytsk3
Pillow
python-magic
matplotlib
reportlab
sqlite3
